Refactor mt_api_test.py
